move key files to lib folder
