+++ lib/error.js
