+++ lib/generate.js
@@ -17,8 +17,8 @@ var Path = require("path");
 var Url = require("url");
 var Util = require("./util");
 
-var TestSectionTpl = fs.readFileSync(__dirname + "/templates/test_section.js.tpl", "utf8");
-var TestHandlerTpl = fs.readFileSync(__dirname + "/templates/test_handler.js.tpl", "utf8");
+var TestSectionTpl = fs.readFileSync(__dirname + "/../templates/test_section.js.tpl", "utf8");
+var TestHandlerTpl = fs.readFileSync(__dirname + "/../templates/test_handler.js.tpl", "utf8");
 
 var main = module.exports = function() {
     // check routes path
@@ -176,7 +176,7 @@ var main = module.exports = function() {
     Util.log("Converting routes to functions");
     prepareApi(routes);
 
-    var apidocsPath = Path.join(__dirname, "doc", "apidoc.js");
+    var apidocsPath = Path.join(__dirname, "/../doc", "apidoc.js");
     fs.writeFileSync(apidocsPath, apidocs);
 
     Object.keys(sections).forEach(function(section) {
@@ -187,7 +187,7 @@ var main = module.exports = function() {
             .replace("<%version%>", "3.0.0")
             .replace(/<%sectionName%>/g, section)
             .replace("<%testBody%>", def.join("\n\n"));
-        var path = Path.join(__dirname, "test", section + "Test.js");
+        var path = Path.join(__dirname, "/../test", section + "Test.js");
 
         fs.writeFileSync(path, body, "utf8");
     });
